split some shared helpers into separate file
